fix: incorrect manage url for real yield eth vault (#712)
The manage url for the RYETH vault for Arbitrum was linking to the RYETH
vault for Optimism (it's quite noticeable because the vaults have
different APY and TVL's). The root cause is that I was using address for
the cellar config key, which is not unique for all vaults. This PR
updates the config to use the same id's as in the somm strategies api.

Related to RET-1258